Use floats not ints for total budget calculation

+++ sites/all/modules/ict_project/js/ict_project.js
@@ -7,8 +7,8 @@
     $('tr.sum input.form-text').keyup(function(){
       var table = $(this).parents('table').attr('id');
       $("#" + table +' td.total').each(function(i){
-        var sum = parseInt($("#" +table+" td.opex[data-sum='" + i + "']").find('.form-text').val())+parseInt($("#" +table+ " td.capex[data-sum='" + i + "']").find('.form-text').val());
-        var total =  (sum === parseInt(sum, 10)) ? sum : '';
+        var sum = parseFloat($("#" +table+" td.opex[data-sum='" + i + "']").find('.form-text').val())+parseFloat($("#" +table+ " td.capex[data-sum='" + i + "']").find('.form-text').val());
+        var total =  (sum === parseFloat(sum, 10)) ? sum : '';
         $(this).find('.form-text').val(total);
       });
